recombined packaging scripts, buildbot argument allows for smooth packaging in buildbot environment

+++ tools/package_standalone.pl
@@ -4,27 +4,39 @@
 #
 use strict;
 use warnings;
-`rm -rf PhyloSift`;
 
-my $branch = $ARGV[0] || "master";
-
-#`rm -rf bioperl-live`;
-`git clone git://github.com/gjospin/PhyloSift.git`;
-`cd PhyloSift ; git checkout $branch`;
-`rm -rf PhyloSift/.git`;
-`rm PhyloSift/Makefile.PL`;
-`rm PhyloSift/ignore.txt`;
-`rm PhyloSift/MANIFEST`;
-`rm PhyloSift/Changes`;
-`rm PhyloSift/.gitignore`;
-`rm PhyloSift/.includepath`;
-`rm PhyloSift/.project`;
-`rm -rf PhyloSift/t`;
-`rm -rf PhyloSift/tools`;
-
-# add bioperl-live
-`git clone git://github.com/bioperl/bioperl-live.git`;
-`mv bioperl-live/Bio* PhyloSift/lib`;
+my $prefix = "";
+if ($ARGV[0] eq "buildbot") # if buildbot
+{
+	$prefix = ".."; # set path prefix
+	
+	# add bioperl-live (working in buildbot)
+    `git clone git://github.com/bioperl/bioperl-live.git`;
+    `mv bioperl-live/Bio* lib/`;
+}else{
+	`rm -rf PhyloSift`;
+	
+	my $branch = $ARGV[0] || "master"; # set branch to checkout
+    $prefix = "../PhyloSift"; # set path prefix
+    
+    #`rm -rf bioperl-live`;
+    `git clone git://github.com/gjospin/PhyloSift.git`;
+    `cd PhyloSift ; git checkout $branch`;
+    `rm -rf PhyloSift/.git`;
+    `rm PhyloSift/Makefile.PL`;
+    `rm PhyloSift/ignore.txt`;
+    `rm PhyloSift/MANIFEST`;
+    `rm PhyloSift/Changes`;
+    `rm PhyloSift/.gitignore`;
+    `rm PhyloSift/.includepath`;
+    `rm PhyloSift/.project`;
+    `rm -rf PhyloSift/t`;
+    `rm -rf PhyloSift/tools`;
+    
+    # add bioperl-live (not on buildbot)
+    `git clone git://github.com/bioperl/bioperl-live.git`;
+    `mv bioperl-live/Bio* PhyloSift/lib/`;
+}
 
 # add Rutger Vos' Bio::Phylo
 `curl -LO http://search.cpan.org/CPAN/authors/id/R/RV/RVOSA/Bio-Phylo-0.45.tar.gz`;
@@ -32,7 +44,7 @@ my $branch = $ARGV[0] || "master";
 chdir("Bio-Phylo-0.45");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/Bio/Phylo* ../PhyloSift/lib/Bio/`;
+`mv blib/lib/Bio/Phylo* $prefix/lib/Bio/`;
 chdir("..");
 
 # add JSON package
@@ -41,7 +53,7 @@ chdir("..");
 chdir("JSON-2.53");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/JSON* ../PhyloSift/lib/`;
+`mv blib/lib/JSON* $prefix/lib/`;
 chdir("..");
 
 # Encode::Locale
@@ -50,7 +62,7 @@ chdir("..");
 chdir("Encode-Locale-0.04");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/Encode ../PhyloSift/lib/`;
+`mv blib/lib/Encode $prefix/lib/`;
 chdir("..");
 
 # add Locale::Maketext
@@ -62,7 +74,7 @@ chdir("Locale-Maketext-1.19");
 `rm lib/Locale/Maketext/*.pod`;
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/Locale/ ../PhyloSift/lib/`;
+`mv blib/lib/Locale/ $prefix/lib/`;
 chdir("..");
 
 # Locale::Maketext::Simple 
@@ -71,7 +83,7 @@ chdir("..");
 chdir("Locale-Maketext-Simple-0.21");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/Locale/Maketext/* ../PhyloSift/lib/Locale/Maketext/`;
+`mv blib/lib/Locale/Maketext/* $prefix/lib/Locale/Maketext/`;
 chdir("..");
 
 # XML::Writer
@@ -80,7 +92,7 @@ chdir("..");
 chdir("XML-Writer-0.615");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/XML/ ../PhyloSift/lib/`;
+`mv blib/lib/XML/ $prefix/lib/`;
 chdir("..");
 
 # libwww-perl (for LWP::Simple)
@@ -89,7 +101,7 @@ chdir("..");
 chdir("libwww-perl-6.04");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/LWP/ ../PhyloSift/lib/`;
+`mv blib/lib/LWP/ $prefix/lib/`;
 chdir("..");
 
 `curl -LO http://search.cpan.org/CPAN/authors/id/G/GA/GAAS/HTTP-Message-6.03.tar.gz`;
@@ -97,7 +109,7 @@ chdir("..");
 chdir("HTTP-Message-6.03");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/HTTP/ ../PhyloSift/lib/`;
+`mv blib/lib/HTTP/ $prefix/lib/`;
 chdir("..");
 
 `curl -LO http://search.cpan.org/CPAN/authors/id/G/GA/GAAS/HTTP-Date-6.02.tar.gz`;
@@ -105,7 +117,7 @@ chdir("..");
 chdir("HTTP-Date-6.02");
 `perl Makefile.PL`;
 `make`;
-`mv blib/lib/HTTP/ ../PhyloSift/lib/`;
+`mv blib/lib/HTTP/ $prefix/lib/`;
 chdir("..");
 
 
@@ -117,7 +129,7 @@ chdir("version-0.95");
 `make`;
 
 # put these in "legacy" because we only want to use them if the perl version is ancient -- including them breaks newer perls
-`mv blib/lib/version* ../PhyloSift/legacy/`;
+`mv blib/lib/version* $prefix/legacy/`;
 
 chdir("..");
 
@@ -131,5 +143,5 @@ $datestr .= $timerval[3];
 `mv PhyloSift phylosift_$datestr`;
 `tar cjf phylosift_$datestr.tar.bz2 phylosift_$datestr`;
 `rm -rf phylosift_$datestr`;
-`echo "phylosift_$datestr" > psversion`;
+#`echo "phylosift_$datestr" > psversion`;
 exit 0;
